first small step toward compiler refactor
This commit is the first small move toward moving the compilaton logic
and ast-dependent transformations in zq/compiler.  This commit moves
proc/compiler to the top-level and merges in the compilation and
optimization code from driver.  This is very step on issue #1162.
